changed location of user status functionality
